Update SeleniumHelloWorld.py
Changed File to now search "TonDone", click on the first link, and verifies the text "The easy way to manage hard work" is on the page. If it is on the page the message "Text successfully found" will be printed in the output, otherwise "Text not found" will be printed.